% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{countReplicate}
\alias{countReplicate}
\title{Count numbers of each members}
\usage{
countReplicate(group.vec)
}
\arguments{
\item{group.vec}{Vector of members.}
}
\value{
vector
}
\description{
Count numbers of each members
}
\examples{
countReplicate(c('a','b','b','c','c','c','a','d','d'))
}
