% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{Counts}
\alias{Counts}
\alias{Counts<-}
\alias{Counts,Enone,character,character-method}
\alias{Counts<-,Enone,character,character,matrix-method}
\title{Accessor for the "counts" slot of Enone object.}
\usage{
Counts(object, slot = c("sample", "spike_in"), method)

Counts(object, slot = c("sample", "spike_in"), method) <- value

\S4method{Counts}{Enone,character,character}(object, slot = c("sample", "spike_in"), method)

\S4method{Counts}{Enone,character,character,matrix}(object, slot = c("sample", "spike_in"), method) <- value
}
\arguments{
\item{object}{Enone.}

\item{slot}{Which slot to get, one of \code{sample} or \code{spike_in}.}

\item{method}{Which counts matrix to get, must be one of the raw or normalized counts matrix presented in the selected slot.}

\item{value}{Raw or normalized counts matrix.}
}
\value{
matrix.
}
\description{
Accessor for the "counts" slot of Enone object.
}
