% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{CreateGroupMatrix}
\alias{CreateGroupMatrix}
\title{Create a matrix for RUVSeq}
\usage{
CreateGroupMatrix(group.vec)
}
\arguments{
\item{group.vec}{A vector indicating membership in a group.}
}
\value{
A matrix.
}
\description{
Create a matrix for RUVSeq
}
\examples{
CreateGroupMatrix(c('a','b','b','c','c','c','a','d','d'))
}
