% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.keepConsRows}
\alias{.keepConsRows}
\title{Keep Consistent Rows in Object}
\usage{
.keepConsRows(object)
}
\arguments{
\item{object}{Enone object}
}
\value{
Updated Enone
}
\description{
Make sure the rows in \code{counts} slot and \code{assay} are consistent.
}
