% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.mean_sd}
\alias{.mean_sd}
\title{Statistics summary (mean and +/- sd)}
\usage{
.mean_sd(x)
}
\arguments{
\item{x}{Value}
}
\value{
Vector of mean and mean +/- sd
}
\description{
Statistics summary (mean and +/- sd)
}
