% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromRUVSeq.R
\name{enRUVg}
\alias{enRUVg}
\title{Remove unwanted variation using control genes}
\usage{
enRUVg(object, log = TRUE, k = 1, tolerance = 1e-08, control.idx, drop = 1)
}
\arguments{
\item{object}{A counts matrix.}

\item{log}{Whether to perform log2-transformation with 1 offset on data matrix,
default: TRUE.}

\item{k}{The number of factors of unwanted variation to be estimated from the data.}

\item{tolerance}{Tolerance in the selection of the number of positive singular
values, i.e., a singular value must be larger than tolerance to be considered positive.}

\item{control.idx}{ID of control genes.}

\item{drop}{The number of singular values to drop in the estimation of
unwanted variation, default drop the first singular value that represent the
difference between enrichment and input}
}
\value{
list contain a matrix of unwanted factors (W) and corrected counts matrix (normalizedCounts).
}
\description{
Adapted from \code{RUVSeq} (https://doi.org/10.1038\%2Fnbt.2931)
}
