% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{FilterLowExprGene}
\alias{FilterLowExprGene}
\title{Filter Low Expressed Genes}
\usage{
FilterLowExprGene(x, group = NULL, min.count = 10)
}
\arguments{
\item{x}{Enone object}

\item{group}{Vector or factor giving group membership for a oneway layout,
if appropriate, default: NULL.}

\item{min.count}{Minimum count required for at least some samples, default: 10.}
}
\value{
updated Enone object
}
\description{
Filter Low Expressed Genes
}
