% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelection.R
\name{GeneSelection}
\alias{GeneSelection}
\title{Gene set selection}
\usage{
GeneSelection(object, n.neg.control = 1000, n.pos.eval = 500, n.neg.eval = 500)
}
\arguments{
\item{object}{Enone object}

\item{n.neg.control}{Number of negative control genes for RUV normalization, default: 1000.}

\item{n.pos.eval}{Number of positive evaluation genes for wanted variation assessment, default: 500.}

\item{n.neg.eval}{Number of negative evaluation genes for unwanted variation assessment, default: 500.}
}
\value{
list of genes
}
\description{
Select negative control genes for RUV normalization, positive
and negative evaluation genes for assessment.
}
