% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{getEnrichment}
\alias{getEnrichment}
\alias{getEnrichment,Enone,character-method}
\title{Accessor of enrichment results}
\usage{
getEnrichment(object, slot = c("sample", "spike_in"), filter = FALSE)

\S4method{getEnrichment}{Enone,character}(object, slot = c("sample", "spike_in"), filter = FALSE)
}
\arguments{
\item{object}{Enone.}

\item{slot}{Which slot to get, one of \code{sample} or \code{spike_in}.}

\item{filter}{Whether to get the filtered enrichment, default FALSE.}
}
\value{
list of enrichment table
}
\description{
Get all or filtered enrichment
}
