% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{getMetrics}
\alias{getMetrics}
\alias{getMetrics,Enone-method}
\title{Accessor of Enone metrics}
\usage{
getMetrics(object)

\S4method{getMetrics}{Enone}(object)
}
\arguments{
\item{object}{Enone.}
}
\value{
data.frame
}
\description{
Accessor of Enone metrics
}
