% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{listParameter}
\alias{listParameter}
\alias{getParameter}
\alias{listParameter,Enone-method}
\alias{getParameter,Enone,character-method}
\title{Accessor of Enone parameter}
\usage{
listParameter(object)

getParameter(object, name)

\S4method{listParameter}{Enone}(object)

\S4method{getParameter}{Enone,character}(object, name)
}
\arguments{
\item{object}{Enone.}

\item{name}{Name of the parameter.}
}
\value{
Vector of parameter names

Vector of parameter
}
\description{
List all parameters.

Get specific parameter.
}
