% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{getScore}
\alias{getScore}
\alias{getScore,Enone-method}
\title{Accessor of Enone score}
\usage{
getScore(object)

\S4method{getScore}{Enone}(object)
}
\arguments{
\item{object}{Enone.}
}
\value{
data.frame
}
\description{
Accessor of Enone score
}
