% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{listNormalization}
\alias{listNormalization}
\alias{listNormalization,Enone-method}
\title{Accessor of normalization methods}
\usage{
listNormalization(object)

\S4method{listNormalization}{Enone}(object)
}
\arguments{
\item{object}{Enone.}
}
\value{
Vector of normalization methods
}
\description{
List all normalization
}
