% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{normPossionSeq}
\alias{normPossionSeq}
\title{Perform PossionSeq normalization}
\usage{
normPossionSeq(data, ...)
}
\arguments{
\item{data}{A un-normalized count data matrix of shape n x p, where n is
the number of samples and p is the number of features.}

\item{...}{Additional parameters can be passed to \code{PS.Est.Depth}.}
}
\value{
List containing normalized counts and normalized factors for library size.
}
\description{
Perform PossionSeq normalization
}
