% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{normTC}
\alias{normTC}
\title{Perform total count normalization}
\usage{
normTC(data)
}
\arguments{
\item{data}{A un-normalized count data matrix of shape n x p, where n is
the number of samples and p is the number of features.}
}
\value{
List containing normalized counts and normalized factors for library size.
}
\description{
Perform total count normalization
}
