% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{OutlierTest}
\alias{OutlierTest}
\title{Outlier Test}
\usage{
OutlierTest(x, return = FALSE, remove = FALSE)
}
\arguments{
\item{x}{Enone object}

\item{return}{Whether to return object or simply perform test, default FALSE}

\item{remove}{Whether to remove outliers. If TRUE, must be paired with
"return=TRUE", default: FALSE}
}
\value{
updated Enone object
}
\description{
Rosner’s outlier test on principal component 1 to assess and
remove potential outliers.
}
