% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{PCA_Biplot}
\alias{PCA_Biplot}
\title{Biplot of individuals and variables}
\usage{
PCA_Biplot(object, score, pt.label = TRUE, interactive = FALSE)
}
\arguments{
\item{object}{PCA object returned by \code{prcomp}.}

\item{score}{Vector of performance scores from \code{Enone} evaluation score.}

\item{pt.label}{Whether to plot the point labels, default: TRUE.}

\item{interactive}{Whether to demonstrate the plot interactively, default: FALSE.}
}
\value{
plot
}
\description{
Biplot of individuals and variables
}
