% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{PCAplot}
\alias{PCAplot}
\title{PCA plot from counts matrix}
\usage{
PCAplot(
  object,
  use.pc = c(1, 2),
  color = NULL,
  label = NULL,
  shape = NULL,
  vst.norm = FALSE,
  palette = NULL,
  repel = TRUE
)
}
\arguments{
\item{object}{A count matrix.}

\item{use.pc}{Which two PCs to be used, default PC1 in x-axis and PC2 in y-axis.}

\item{color}{Vector indicates the color mapping of samples, default NULL.}

\item{label}{Vector of sample names or labels, default NULL.}

\item{shape}{Vector indicates the shape mapping of samples, default NULL.}

\item{vst.norm}{Whether to perform \code{vst} transformation, default FALSE.}

\item{palette}{The color palette for different groups.}

\item{repel}{Whether to use \code{ggrepel} to avoid overlapping text labels or not, default TRUE.}
}
\value{
ggplot2 object
}
\description{
PCA plot from counts matrix
}
\details{
Perform PCA based on matrix using \code{prcomp}, and visualized with scatter plot.
}
