% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromPossionSeq.R
\name{PS.Est.Depth}
\alias{PS.Est.Depth}
\title{Esimate sequencing depth}
\usage{
PS.Est.Depth(n, iter = 5, ct.sum = 5, ct.mean = 0.5)
}
\arguments{
\item{n}{The data matrix. The rows are counts for a gene, and the columns are counts from an experiment.}

\item{iter}{Number of iterations used. Default value: 5. The default value is usually a good choice.}

\item{ct.sum}{if the total number of reads of a gene across all experiments <= ct.sum, this gene will not be considered for estimating sequencing depth. Default value: 5.}

\item{ct.mean}{if the mean number of reads of a gene across all experiments <= ct.mean, this gene will not be considered for estimating sequencing depth. Default value: 0.5.}
}
\value{
estimated sequencing depth. a vector. their product is 1.
}
\description{
Functions from PossionSeq (https://doi.org/10.1093/biostatistics/kxr031)
}
\examples{
mat <- matrix(rnbinom(300, mu=100, size=1), ncol=5)
PS.Est.Depth(mat)

}
