% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{synEnrichment}
\alias{synEnrichment}
\title{Enrichment level of synthetic RNA}
\usage{
synEnrichment(object, method = "TC")
}
\arguments{
\item{object}{Enone object.}

\item{method}{Which normalization methods to perform.}
}
\value{
enrichment level of synthetic RNA
}
\description{
Enrichment level of synthetic RNA
}
