% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{UseNormalization}
\alias{UseNormalization}
\title{Apply specific normalization method}
\usage{
UseNormalization(object, slot = c("sample", "spike_in"), method)
}
\arguments{
\item{object}{Enone object.}

\item{slot}{Which slot, one of \code{sample} or \code{spike_in}.}

\item{method}{Which normalization methods to perform.}
}
\value{
object
}
\description{
Apply specific normalization method
}
