% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{BetweenStatPlot}
\alias{BetweenStatPlot}
\title{Box-violin plot comparing values between groups}
\usage{
BetweenStatPlot(
  data,
  x,
  y,
  color,
  palette = NULL,
  test = c("wilcox.test", "t.test", "none"),
  comparisons = NULL,
  step.increase = 0.3
)
}
\arguments{
\item{data}{A data frame (or a tibble).}

\item{x}{The grouping variable from the \code{data}.}

\item{y}{The value variable from the \code{data}.}

\item{color}{The color variable from the \code{data}.}

\item{palette}{The color palette for different groups.}

\item{test}{Perform "wilcox.test" or "t.test" or no test.}

\item{comparisons}{A list of length-2 vectors specifying the groups of interest to be compared. For example to compare groups "A" vs "B" and "B" vs "C", the argument is as follow: comparisons = list(c("A", "B"), c("B", "C"))}

\item{step.increase}{numeric vector with the increase in fraction of total height for every additional comparison to minimize overlap.}
}
\value{
ggplot2 object
}
\description{
Box-violin plot comparing values between groups
}
