% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{getGeneSet}
\alias{getGeneSet}
\alias{getGeneSet,Enone,character-method}
\title{Accessor of gene set}
\usage{
getGeneSet(object, name)

\S4method{getGeneSet}{Enone,character}(object, name)
}
\arguments{
\item{object}{Enone.}

\item{name}{Name of the gene set, must be the same as the column names of \code{rowData(Enone)}.}
}
\value{
Vector of gene ID
}
\description{
Get gene set.
}
