% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{OutlierTest}
\alias{OutlierTest}
\title{Outlier Test}
\usage{
OutlierTest(x, remove = FALSE)
}
\arguments{
\item{x}{Enone object}

\item{remove}{Whether to remove outliers, default: FALSE}
}
\value{
updated Enone object
}
\description{
Rosner’s outlier test on principal component 1 to assess and
remove potential outliers.
}
