% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{reduceRes}
\alias{reduceRes}
\title{Combine list of DE results}
\usage{
reduceRes(res.ls, fc.col, levels = names(res.ls))
}
\arguments{
\item{res.ls}{Named list of differential analysis results tables.
Each elements in the list correspond to a table of differential analysis
results between two groups of samples.}

\item{fc.col}{Name of the fold change column. The fold change column name
must be consistent with your results tables.}

\item{levels}{Factor levels of the groups, default order by the element order of \code{res.ls}.}
}
\value{
data.frame
}
\description{
Combine list of DE results
}
