% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{synEnrichment}
\alias{synEnrichment}
\title{Enrichment level of synthetic RNA}
\usage{
synEnrichment(object, method = "TC", log = TRUE)
}
\arguments{
\item{object}{Enone object.}

\item{method}{Which normalization methods to perform.}

\item{log}{Whether to return \code{log2} values, default: TRUE.}
}
\value{
enrichment level of synthetic RNA
}
\description{
Enrichment level of synthetic RNA
}
